requirements.txt
tensorflow-gpu==2.3.1
tensorflow_datasets==2.0.0
tensorflow-addons==0.8.3
matplotlib==3.1.0
sklearn
seaborn
pandas
opencv-python
setuptools>=41.0.0
Pillow